Add handling of changelog changes
