+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "yarn workspace @macrostrat/storybook run dev",
     "build": "yarn workspace @macrostrat/storybook run build",
+    "status": "tsx scripts/status.ts",
     "prepare": "tsx scripts/prepare.ts",
     "publish": "node scripts/publish.ts"
   },
@@ -34,6 +35,8 @@
     "geologic-patterns": "^2.0.0-dev1",
     "husky": "^4.3.5",
     "lerna": "^4.0.0",
+    "marked": "^11.0",
+    "marked-terminal": "^7.1.0",
     "node-fetch": "^3.2.9",
     "prettier": "^2.7.1",
     "sass-embedded": "^1.79.4",
